fix: preview panel cleanup — remove dual rendering, fix terminal clear, deduplicate file preview code
- Remove legacy FilePreview lightbox overlay that rendered on top of the
  new inline preview (Bug 1)
- Add CLEAR_TERMINAL reducer action so closing the terminal tab actually
  clears its lines instead of appending an unhandled event (Bug 2)
- Fix blob URL leak in handleDownload by revoking after click (Bug 3)
- Extract useFileContent hook and FileContentRenderer component to
  eliminate ~80% code duplication between FilePreviewInline and
  FullscreenPreview (Issue 4)
- Prevent text selection during SplitHandle drag (Issue 5)
- Define --text-secondary CSS variable in global.css and replace
  hardcoded #999 colors across preview CSS modules (Issue 6)
