+++ server/ui/src/components/FullscreenPreview.module.css
@@ -62,7 +62,7 @@
 /* Filename in center */
 .filename {
     font-size: 13px;
-    color: #999; /* var(--text-secondary) */
+    color: var(--text-secondary);
     max-width: 300px;
     overflow: hidden;
     text-overflow: ellipsis;
@@ -87,7 +87,7 @@
     font-size: 12px;
     border: none;
     background: transparent;
-    color: #999;
+    color: var(--text-secondary);
     cursor: pointer;
     border-radius: 3px;
     transition: background 0.15s ease, color 0.15s ease;
@@ -112,7 +112,7 @@
     background: transparent;
     border: none;
     border-radius: 4px;
-    color: #999;
+    color: var(--text-secondary);
     cursor: pointer;
     transition: background 0.15s ease, color 0.15s ease;
 }
@@ -237,7 +237,7 @@
 .statusText {
     padding: 40px;
     text-align: center;
-    color: #999;
+    color: var(--text-secondary);
     font-size: 14px;
 }
 
@@ -258,10 +258,6 @@
 
 :global(body:not(.dark-theme)) .backBtn:hover {
     background: #e5e5e5;
-}
-
-:global(body:not(.dark-theme)) .filename {
-    color: #666;
 }
 
 :global(body:not(.dark-theme)) .toggle {
