+++ server/ui/src/components/PreviewPanel.module.css
@@ -41,7 +41,7 @@
     background: transparent;
     border: none;
     border-bottom: 2px solid transparent;
-    color: #999; /* var(--text-secondary) */
+    color: var(--text-secondary);
     font-size: 13px;
     cursor: pointer;
     white-space: nowrap;
@@ -114,7 +114,7 @@
     background: transparent;
     border: none;
     border-radius: 4px;
-    color: #999;
+    color: var(--text-secondary);
     cursor: pointer;
     transition: background 0.15s ease, color 0.15s ease;
 }
@@ -139,10 +139,6 @@
     border-bottom-color: #ddd;
 }
 
-:global(body:not(.dark-theme)) .tab {
-    color: #666;
-}
-
 :global(body:not(.dark-theme)) .tab:hover {
     background: rgba(0, 0, 0, 0.04);
 }
@@ -158,10 +154,6 @@
 :global(body:not(.dark-theme)) .tabClose:hover {
     background: rgba(0, 0, 0, 0.06);
     color: #222;
-}
-
-:global(body:not(.dark-theme)) .actionBtn {
-    color: #666;
 }
 
 :global(body:not(.dark-theme)) .actionBtn:hover {
